Cleanup BinaryMapper.cs tests and fixed MessageFetchRequest.cs value field type.

+++ Iggy_SDK/Contracts/Http/MessageFetchRequest.cs
@@ -10,7 +10,7 @@ public sealed class MessageFetchRequest
 	public required int ConsumerId { get; init; }
 	public required int PartitionId { get; init; }
 	public required MessagePolling PollingStrategy { get; init; }
-	public required int Value { get; init; }
+	public required ulong Value { get; init; }
 	public required int Count { get; init; }
 	public required bool AutoCommit { get; init; }
 }
